requirements.txt
discord.py
python-dotenv
eth-ape
pre-commit
